♻️ split up foirequest api views
