+++ froide/foirequest/apps.py
@@ -23,11 +23,9 @@ class FoiRequestConfig(AppConfig):
         from froide.account.export import registry
         from froide.api import api_router
         from froide.foirequest import signals  # noqa
-        from froide.foirequest.api_views import (
-            FoiAttachmentViewSet,
-            FoiMessageViewSet,
-            FoiRequestViewSet,
-        )
+        from froide.foirequest.api_views.attachment import FoiAttachmentViewSet
+        from froide.foirequest.api_views.message import FoiMessageViewSet
+        from froide.foirequest.api_views.request import FoiRequestViewSet
         from froide.helper.search import search_registry
         from froide.team import team_changed
